refactor(orders): B2B-4615 extract pagination into useUnifiedOrdersPagination hook
Extract cursor pagination state and logic into a dedicated
useUnifiedOrdersPagination hook, composed inside
useCustomerOrdersFilterState. Keeps filter/sort concerns separate
from pagination. Reverts the file rename — useCustomerOrdersFilterState
keeps its original name from B2B-4614.
